.history/requirements_20251208225620.txt
fastapi
uvicorn[standard]
openpyxl
itsdangerous
eportlab